requirements.txt
google-generativeai
reportlab
svglib
matplotlib
seaborn
pandas